use findfirstvisibleitem for multiselectmode activation and findlastvisibleitem for forcing reload
last commit; going to do a refactor to get rid of groupie and building a nested recyclerview (so I can actually diffutil properly) myself. Groupie does not have enough documentation to show me properly how to use updateAsync (diffutil) with expandables and my attempts of making a group and calling updateAsync only got as far as the shoppinglist blinking and disappearing. Instead of wasting more days and not getting anywhere I'm ditching groupie.